implementing and update all delete methods with cascade constraintes

+++ stock-de-medaille/src/main/java/bf/gov/gcob/medaille/services/ServiceImpl/DistinctionServiceImpl.java
@@ -3,13 +3,16 @@ package bf.gov.gcob.medaille.services.ServiceImpl;
 import bf.gov.gcob.medaille.mapper.DistinctionMapper;
 import bf.gov.gcob.medaille.model.dto.DistinctionDTO;
 import bf.gov.gcob.medaille.model.entities.Distinction;
+import bf.gov.gcob.medaille.model.entities.Medaille;
 import bf.gov.gcob.medaille.repository.DistinctionRepository;
+import bf.gov.gcob.medaille.repository.MedailleRepository;
 import bf.gov.gcob.medaille.services.DistinctionService;
 import java.util.List;
 import java.util.stream.Collectors;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import org.springframework.util.CollectionUtils;
 
 @Slf4j
 @Service
@@ -18,6 +21,7 @@ public class DistinctionServiceImpl implements DistinctionService {
 
     private final DistinctionMapper distinctionMapper;
     private final DistinctionRepository distinctionRepository;
+    private final MedailleRepository medailleRepository;
 
     @Override
     public DistinctionDTO create(DistinctionDTO distinctionDTO) {
@@ -26,14 +30,14 @@ public class DistinctionServiceImpl implements DistinctionService {
         distinctionDTO = distinctionMapper.toDTO(distinction);
         return distinctionDTO;
     }
-    
+
     @Override
     public List<DistinctionDTO> findAll() {
         log.info("Liste de toutes les distinctions");
         List<DistinctionDTO> distinctions = distinctionRepository.findAll().stream().map(distinctionMapper::toDTO).collect(Collectors.toList());
         return distinctions;
     }
-    
+
     @Override
     public DistinctionDTO update(DistinctionDTO distinctionDTO) {
         Distinction distinction = distinctionRepository.findById(distinctionDTO.getIdDistinction()).orElse(null);
@@ -45,9 +49,15 @@ public class DistinctionServiceImpl implements DistinctionService {
         DistinctionDTO ordonnateurModifer = distinctionMapper.toDTO(distinction);
         return ordonnateurModifer;
     }
-    
+
     @Override
     public void delete(Long idDistinction) {
-        distinctionRepository.deleteById(idDistinction);
+        log.info("Suppression de la distinction : {}", idDistinction);
+        List<Medaille> medailles = medailleRepository.findByDistinctionIdDistinction(idDistinction);
+        if (medailles != null || !CollectionUtils.isEmpty(medailles)) {
+            throw new RuntimeException("Veuillez supprimer les medailles... de cette distinction avant de poursuivre.");
+        } else {
+            distinctionRepository.deleteById(idDistinction);
+        }
     }
 }
